+++ src/main/java/io/github/fablabsmc/fablabs/api/fluidvolume/v1/volume/api/SingleFluidContainer.java
@@ -7,7 +7,6 @@ import io.github.fablabsmc.fablabs.api.fluidvolume.v1.math.Fraction;
 
 import net.minecraft.fluid.Fluid;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.text.LiteralText;
 import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.Identifier;
@@ -19,8 +18,7 @@ import net.minecraft.util.registry.Registry;
 public interface SingleFluidContainer extends FluidContainer {
 	/**
 	 * get the fluid this volume contains.
-	 * if the fluid is EMPTY, then {@link #getTotalVolume()}
-	 * must be zero, a EMPTY fluid represents no fluid at all
+	 * if the fluid is EMPTY, then {@link #getTotalVolume()} must be zero, a EMPTY fluid represents no fluid at all.
 	 *
 	 * @return the fluid this volume represents
 	 */
@@ -37,11 +35,13 @@ public interface SingleFluidContainer extends FluidContainer {
 	default Text toText() {
 		Fraction total = this.getTotalVolume();
 		Identifier fluid = Registry.FLUID.getId(this.getFluid());
-		if(total.equals(ONE))
-			return new TranslatableText("text.fluid.singular", total, new TranslatableText("fluid."+fluid.getNamespace()+"."+fluid.getPath()));
-		else if(total.equals(ZERO))
+
+		if (total.equals(ONE)) {
+			return new TranslatableText("text.fluid.singular", total, new TranslatableText("fluid." + fluid.getNamespace() + "." + fluid.getPath()));
+		} else if (total.equals(ZERO)) {
 			return new TranslatableText("text.fluid.empty");
-		else
-			return new TranslatableText("text.fluid.plural", new TranslatableText("fluid."+fluid.getNamespace()+"."+fluid.getPath()));
+		} else {
+			return new TranslatableText("text.fluid.plural", new TranslatableText("fluid." + fluid.getNamespace() + "." + fluid.getPath()));
+		}
 	}
 }
